Add app file with the express server and add data base into server file
